reorganize game page and update game keyboard
